pyproject.toml
[tool.poetry]
name = "container-mngr"
version = "0.2.0"
description = "Simple console application for managing Docker containers"
authors = ["Stefan Szarek <stefan.szarek@gmail.com>"]
readme = "README.md"
packages = [{include = "src/container_mngr"}]

[tool.poetry.dependencies]
python = "^3.10"
textual = "^0.5.0"
docker = "^6.0.1"
pypiwin32 = {version = "^223", platform = "win32"}
python-dateutil = "^2.8.2"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^22.10.0"
pytest = "^7.2.0"
mypy = "^0.991"
commitizen = "^2.37.1"
pre-commit = "^2.20.0"
syrupy = "^3.0.5"
pytest-asyncio = "^0.20.3"
pytest-cov = "^4.0.0"
types-python-dateutil = "^2.8.19.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
includes = "src"

[tool.commitizen]
version = "0.2.0"
update_changelog_on_bump = true
annotated_tag = true
version_files = [
  "pyproject.toml:version",
  "src/container_mngr/__version__.py"
]